requirements.txt
sqlalchemy
fastapi[standard]
bcrypt
pyjwt
python-dotenv
python-jose